chore: Comment out flatpak service in configuration.nix
This commit comments out the flatpak service in the configuration.nix file. This change is in line with recent user commits that have updated the system packages and enabled services. By commenting out the flatpak service, it ensures that it is not enabled in the system. This change aligns with the goal of improving the functionality and compatibility of the system.